Create README.md
This was a learn the basics of Bootstrap offered by Codecademy to code a fictional website for the Adhoc Ad agency. The assignment was to give the employees a carousel and the tenets of the company cards.  To challenge myself I flipped the project and made the tenets carousel and the employees cards.  I added copy to the basic project.